myvenv/easymedico/requirements.txt
astroid==2.4.2
cffi==1.14.2
colorama==0.4.3
cryptography==3.1
Django==2.2.16
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
isort==5.5.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.2.0
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wfastcgi==3.0.0
wrapt==1.12.1
